+++ pypsa/optimization/abstract.py
@@ -349,8 +349,13 @@ def optimize_mga(
         Keyword argument used by `linopy.Model.solve`, such as `solver_name`,
 
     Returns
-    -------
-    None
+    status : str
+        The status of the optimization, either "ok" or one of the codes listed
+        in https://linopy.readthedocs.io/en/latest/generated/linopy.constants.SolverStatus.html
+    condition : str
+        The termination condition of the optimization, either
+        "optimal" or one of the codes listed in
+        https://linopy.readthedocs.io/en/latest/generated/linopy.constants.TerminationCondition.html
     """
     if snapshots is None:
         snapshots = n.snapshots
@@ -424,7 +429,7 @@ def optimize_mga(
 
     m.objective = merge(objective)
 
-    n.optimize.solve_model(**kwargs)
+    status, condition = n.optimize.solve_model(**kwargs)
 
     # write MGA coefficients into metadata
     n.meta["slack"] = slack
@@ -432,10 +437,12 @@ def optimize_mga(
 
     def convert_to_dict(obj):
         if isinstance(obj, (pd.Series, pd.DataFrame)):
-            return obj.to_dict()
+            return obj.to_dict(orient="list")
         elif isinstance(obj, dict):
             return {k: convert_to_dict(v) for k, v in obj.items()}
         else:
             return obj
 
     n.meta["weights"] = convert_to_dict(weights)
+
+    return status, condition